fix(faculty): fix Faculty Detail layout/positioning bugs, lock term reassignment, credit Clinical Shift duty in workload
Faculty Detail (Courses tab):
- Bring .tab-content/content-card padding in line with the project's 16px rule
  instead of copying student/enquiry/admission-detail's inconsistent 20-24px.
- Add missing section headings (Term Capacity, Assigned Courses, Lab Sessions)
  and reorder/relocate the Term Capacity stats + Raise Cap button onto one row.
- Fix Academic Year/Term selectors stacking vertically: .workload-toolbar's
  local override never reset flex-direction, so the global .content-card rule's
  flex-direction:column cascaded through per-property.
- Fix the Term dropdown showing the raw id after switching Academic Year:
  cms-infinite-select had no resolveLabel/reloadKey, so it couldn't resolve a
  newly-preselected id against its stale cached options.

Raise Cap flyout (shared/flyout-panel, used by 18 screens):
- Root-caused the flyout scrolling with the page instead of staying pinned:
  .detail-page's cms-rise-up entrance animation leaves a computed
  transform:matrix(1,0,0,1,0,0) (never the literal `none`) after it settles,
  which creates a new containing block for position:fixed descendants.
  Rewrote CmsFlyoutPanelComponent to render via CDK Overlay/TemplatePortal
  instead, same pattern already used in shared/column-picker for an
  analogous bug -- escapes the DOM subtree entirely and adds proper
  background-scroll locking while open.

Locked-term guard:
- CourseOfferingSectionFacultyService.upsert/upsertForCohort and
  BatchService.updateBatch/reassignCoordinator now reject on a LOCKED term,
  matching the same requireTermNotLocked pattern already enforced in
  CourseRegistrationServiceImpl/TimetableGenerationService/TermInstanceService.

Clinical Shift duty in workload/capacity:
- Root-caused Faculty Detail's Real Weekly Schedule and Lab Schedules tab
  reading near-zero for faculty coordinating Clinical Shift batches: that
  duty never produces a real ClassSchedule row (confirmed against real data
  and the "fix missing Clinical occurrences" commit two days prior, which
  fixed the identical gap for Timetable Date-wise/Day but missed these two
  Faculty Detail endpoints). ClassScheduleService now merges in
  TimetableSkeletonService's synthetic Clinical Shift grid entries.
- FacultyWorkloadCapacityService's Capacity Planner report had the same gap
  in its committed-hours/session-count figures -- fixed the same way, plus
  a new CurriculumHoursCalculator.clinicalShiftPeriodsEquivalent for the
  fractional periods-equivalent (e.g. 370min/50min-period = 7.4, not rounded).
- Both use the shift's raw clinicalShiftDurationMinutes, not the displayed
  bus-inclusive window (which adds travel buffer each way), so committed
  hours stay on the same footing as curriculum-derived demand.

+++ backend/src/main/java/com/cms/dto/FacultyWorkloadRow.java
@@ -16,8 +16,12 @@ public record FacultyWorkloadRow(
     boolean overCommitted,
     /** Real session (period-row) count this faculty currently has this week -- the same unit
      *  {@code effectiveMinSessions} is configured in, so this and the floor are always directly
-     *  comparable, unlike committedHoursPerWeek (variable-length periods). */
-    int actualSessionsPerWeek,
+     *  comparable, unlike committedHoursPerWeek (variable-length periods). Fractional since a
+     *  faculty coordinating Clinical Shift duty (never a real period-row -- see
+     *  FacultyWorkloadCapacityService) contributes a periods-equivalent derived from the shift's
+     *  real-world duration, which rarely lines up with a whole number of periods (e.g. 7.4). A
+     *  real ClassSchedule row always contributes exactly 1.0. */
+    double actualSessionsPerWeek,
     boolean minSessionsConfigured,
     Integer effectiveMinSessions,
     /** Advisory only -- never blocks anything, unlike overCommitted's hard-cap counterpart. */
